Moved fragments endpoint to client RPC provider
